+++ src/const.ts
@@ -1,3 +1,3 @@
 export const LOGIN_PATH = "/login";
 
-export const APP_VERSION = "0.9.4";
+export const APP_VERSION = "0.9.5";